backend/requirements.txt
Django==4.2.0
django-cors-headers==4.0.0
django-filter==23.2
django-debug-toolbar==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow>=10.0.0
python-dotenv==1.0.0
python-dateutil==2.8.2
requests==2.28.0
gunicorn==20.1.0
markdown==3.4.3
whitenoise==6.4.0